Added SimpleForm gem
Simple_form is redundant with Formtastic, which is built into the Surveyor gemset,
however, the application from which surveyor_gui borrows code used Simple_form.
As a matter of expediency, we are using some of the views relying on simple_form
to get Surveyor_gui up and running quickly.  Later refactoring will likely remove
it (and possibly server side views in favor of a javascript client framework like
Ember).

	Gemfile: added simple_form gem
	simple_form.rb: part of simple_form initialization
	simple_form.en.yml: part of simple_form initialization