Override default Surveyor validation on Dependency model
Surveyor validates the Dependency model for the numericality of
question_id and question_group_id.  This validation does not appear
to take place until a corresponding Question has already been committed,
so these id's are available.

However, at some point in the past, we needed to accept_nested_attributes
in the Question model for dependencies.  It's not clear whether this is
still needed, but for the moment, we're keeping it. This means
that the dependency validations are triggered when questions are saved.
And that means the numericality_of validations on Dependency.question_id
should not occur on new records.

	dependency_methods.rb: override validation for numericality_of on
	question_id and question_group_id.  Only do if not a new record.