Fixed the lingering prefix, suffix error (finally, I think).
Extracting the Question and Answer models (which were extensions to the
original Surveyor models) to a module, in the style of Surveyor, broke
the setter for answers.text.  In the case of a number, the setter was
supposed to look at the number prefix and suffix, and use those to set
the text field (the prefix followed by "|" followed by sufix).

However, because the accepts_nested_attributes call was now being set
using base.send, the prefix and suffix setters, which changed answer.first.text,
were being overrode by the text setter created by accepts_nested_attributes.  This
was the opposite of the required behaviour.

Tried creating a text setter method in the module, but that didn't work either.
It never got called.

The solution was to create an alternative text setter using the class_eval call on
base.  The new text setter successfully overrides the default text setter created by
accepts_nested_attributes.

The new text setter only set the text if the question_type is not number.  If the question_type
is number, the prefix and suffix setters do the work, and are not stepped on by the text setter.

	app/controllers/questions_controller.rb: make sure to set text back to original_choice when switching to a multiple
	choice (pick) question - changed render_picks_partial.  Uses update_attribute against the text field.

	app/views/questions/_no_picks.html.erb: minor change to format the width of the suffix/prefix fields.  Got stepped on
	somewhere back in the commit trail.

	lib/assets/stylesheets/surveyor_gui.sass: part of the formatting change for suffix/prefix.

	lib/surveyor_gui/models/answer_methods.rb: reverted to an older version that mirrors the original code taken from
	evaluationrx.com.

	lib/surveyor_gui/models/question_methods.rb: use class_eval to set a new text setter on base.  Restore
	setters for prefix and suffix