Dependency editor bug - fixed problem adding more than one row.
Found that when entering more than one row, all but one is squashed.
That's because the attribute numbering is stuck at 0 for all new rows.
Added call to the replace_idx method in surveyor_gui.js.  This will
renumber all attributes, which is necessary for the update to correctly
process the parameters.

Need to do same thing on add answers in the question editor.

Also changed the dependencies controller to correct numbering on the
questions when label questions are present.

	app/controllers/dependencys_controller.rb: fixed the numbering on
	dependency questions.

	app/views/dependencys/_dependency_condition_fields.html.erb:
	added dependency_condition.id hidden field.

	app/views/surveyforms/_question_fields.html.erb: changed the
	width of the dependency_conditions editor.

	lib/assets/javascripts/surveyor_gui/surveyor_gui.js: added call
	to replace_idx when rendering dependency_conditions_partial.