Temporary fix for error in updating dependencies.
Refactoring caused update dependencies to break.  Saving question
was not saving nested attributes, so dependency and dependency condition were
not getting updated.

The culprit was a new subclass of question added to QuestionMethods.
When this class was moved to the dependency controllers instead, it worked.

This is a bad solution, but I will revisit after studying metaprogramming in
greater depth.

	spec/controllers/surveyor_controller_spec.rb