Fixed duplicate grid question mentioned in Issue 7
Was not correctly entering the starting display order.  Added
parameter to TextboxToRecord initializer.

This also fixes the issue of grid questions being placed in the wrong
spot.

Also noticed that delete was not cleaning up spare rows in the grid.
Question has a belongs_to dependent => :destroy on QuestionGroup, so if you delete
the question, the question group goes.  But the QuestionGroup does not
have a corresponding dependent => :destroy, so a deleted QuestionGroup may
have orphans.  Added cleanup to the QuestionController destroy method to
allow for this.

	modified:   app/controllers/questions_controller.rb
	modified:   lib/surveyor_gui/models/question_methods.rb