Grid dropdowns working :tada:
Got rid of "omit".  Doesn't seem to fit this format.
Fixed other, although it does not
add a text box when selected.  Could be made to do this with js, but
corresponding magic would be needed on Surveyor.  Not sure its
necessary.

Added cleanup when switching from grid_dropdown to grid_any/grid_other -
needed to delete orphan columns.  Not needed for other question_types,
since group will be deleted and take columns with it.

where clause on answers not picking up nil = null, so added "or IS NULL".
Creating dups otherwise.
	modified:   app/views/questions/_grid_dropdown_fields.html.erb -
	removed call to render _answer_options partial and replaced it with
	a skinnier version of the _answer_options, one that does not include
	omit.

	modified:   lib/surveyor_gui/models/question_type_methods.rb