Ensure that the dependency editor renders the question pulldown correctly for groups.
When adding dependencies to a group question, the dependency editor was (correctly)
excluding the first question in the group from the question dropdown, but was
(incorrectly) showing the subsequent questions that belonged to the group.

The group needs to be treated as a single question, and this change ensures
no questions in a group are shown in the question dropdown when adding dependencies
to that group.
	modified:   app/controllers/dependencys_controller.rb